# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WaytoAlgo API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/profile-setup:
    post:
      operationId: setupProfile
      tags: [auth]
      summary: Complete profile setup after registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileSetupBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
        country:
          type: string
        referralCode:
          type: string
      required: [name, email, phone, password, country]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    ProfileSetupBody:
      type: object
      properties:
        walletAddress:
          type: string
        country:
          type: string
        idNumber:
          type: string
        profileImage:
          type: string
      required: [walletAddress, country]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        referralCode:
          type: string
        sponsorId:
          type: integer
          nullable: true
        walletAddress:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        profileImage:
          type: string
          nullable: true
        currentLevel:
          type: integer
        currentRankId:
          type: integer
          nullable: true
        isAdmin:
          type: boolean
        isActive:
          type: boolean
        withdrawalBlocked:
          type: boolean
        p2pBlocked:
          type: boolean
        investmentBlocked:
          type: boolean
        blockReason:
          type: string
          nullable: true
        walletBalance:
          type: number
        hyperCoinBalance:
          type: number
        profileComplete:
          type: boolean
        totalEarnings:
          type: number
        totalInvested:
          type: number
        createdAt:
          type: string
      required: [id, name, email, phone, referralCode, currentLevel, isAdmin, isActive, withdrawalBlocked, p2pBlocked, investmentBlocked, walletBalance, hyperCoinBalance, profileComplete, totalEarnings, totalInvested, createdAt]
...